common/requirements.txt
xmltodict
bson
sqlalchemy
paramiko
python-magic
